♻️ (aa-devtool-saas): Split signature module into per-provider verifiers
AAASM-924 ACs require that each provider's verifier live in its own
module and that no scheme be shared across providers. Convert
`signature.rs` into a `signature/` folder with `claude_ai.rs`,
`chatgpt.rs`, and `cursor.rs`, each owning its own header constant,
verification body, and tests. `signature/mod.rs` retains the shared
`SignatureError` and a thin dispatcher in `verify`.
